jiwq: Add shared work queue driver

+++ jiwq.c
@@ -49,13 +49,25 @@ static struct jiwq_driver {
 		.default_retry_nr	= 5,	/* 5 retries */
 		.default_delay_ms	= 0,	/* no delay */
 		.proc			= NULL,
-		.name			= "jiwqsingle",
+		.name			= "jisinglewq",
 	},
 	{
 		.default_retry_nr	= 5,	/* 5 retries */
 		.default_delay_ms	= 0,	/* no delay */
 		.proc			= NULL,
-		.name			= "jiwqsingledelay",
+		.name			= "jisinglewqdelay",
+	},
+	{
+		.default_retry_nr	= 5,	/* 5 retries */
+		.default_delay_ms	= 0,	/* no delay */
+		.proc			= NULL,
+		.name			= "jisharedwq",
+	},
+	{
+		.default_retry_nr	= 5,	/* 5 retries */
+		.default_delay_ms	= 0,	/* no delay */
+		.proc			= NULL,
+		.name			= "jisharedwqdelay",
 	},
 };
 
@@ -156,7 +168,9 @@ static int __init init(void)
 			end = drv;
 			goto err;
 		}
-		if (!strncmp(drv->name, "jiwqsingle", strlen("jiwqsingle")))
+		if (!strncmp(drv->name, "jisharedwq", strlen("jisharedwq")))
+			wq = system_wq;
+		else if (!strncmp(drv->name, "jisinglewq", strlen("jisinglewq")))
 			wq = create_singlethread_workqueue(drv->name);
 		else
 			wq = create_workqueue(drv->name);
@@ -175,7 +189,8 @@ static int __init init(void)
 		fops->release	= seq_release;
 		proc = proc_create_data(path, S_IWUSR|S_IRUGO, NULL, fops, drv);
 		if (IS_ERR(proc)) {
-			destroy_workqueue(wq);
+			if (drv->wq != system_wq);
+			destroy_workqueue(drv->wq);
 			err = PTR_ERR(proc);
 			end = drv;
 			goto err;
@@ -185,7 +200,8 @@ static int __init init(void)
 	return 0;
 err:
 	for (drv = jiwq_drivers; drv != end; drv++) {
-		destroy_workqueue(drv->wq);
+		if (drv->wq != system_wq)
+			destroy_workqueue(drv->wq);
 		proc_remove(drv->proc);
 	}
 	return err;
@@ -198,7 +214,8 @@ static void __exit term(void)
 	struct jiwq_driver *end = drv+ARRAY_SIZE(jiwq_drivers);
 
 	for (drv = jiwq_drivers; drv != end; drv++) {
-		destroy_workqueue(drv->wq);
+		if (drv->wq != system_wq)
+			destroy_workqueue(drv->wq);
 		proc_remove(drv->proc);
 	}
 }
